Fixes to AttentionDetector.__init__
`self.model.query` raises due to the unexpected `return_type` argument.
I believe `inference` was intended here, since this is how `detect`
operates.

Also when both `pos_examples` and `neg_examples` are provided,
`__init__` ignored the positive scores entirely and uses the mean
negative score as a threshold, resulting in a ~50% FPR, which surely
isn't intended. I replace this with the average of the mean positive and
negative score to distinguish the two groups; this is probably not
optimal but is the closest reasonable option to what was previously
used.